requirements.txt
fastapi
uvicorn
pydantic
numpy
opencv-python
ultralytics
mediapipe
spotipy
python-dotenv
